[240802][리뷰 구현]

+++ src/main/java/com/example/marketReservation/service/ReviewService.java
@@ -3,9 +3,7 @@ package com.example.marketReservation.service;
 import com.example.marketReservation.MarketReservationApplication;
 import com.example.marketReservation.domain.MarketEntity;
 import com.example.marketReservation.domain.MemberEntity;
-import com.example.marketReservation.domain.ReservationEntity;
 import com.example.marketReservation.domain.ReviewEntity;
-import com.example.marketReservation.model.Reservation;
 import com.example.marketReservation.model.Review;
 import com.example.marketReservation.repository.MarketRepository;
 import com.example.marketReservation.repository.MemberRepository;
@@ -16,7 +14,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.lang.reflect.Member;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -31,11 +28,19 @@ public class ReviewService {
 
     final private MemberRepository memberRepository;
 
-    public ReviewService(ReviewRepository reviewRepository, ReservationRepository reservationRepository, MarketRepository marketRepository, MemberRepository memberRepository) {
+    final private UserService userService;
+
+    public ReviewService(ReviewRepository reviewRepository,
+                         ReservationRepository reservationRepository,
+                         MarketRepository marketRepository,
+                         MemberRepository memberRepository,
+                         UserService userService)
+    {
         this.reviewRepository = reviewRepository;
         this.reservationRepository = reservationRepository;
         this.marketRepository = marketRepository;
         this.memberRepository = memberRepository;
+        this.userService = userService;
     }
 
     final private Logger logger = LoggerFactory.getLogger(MarketReservationApplication.class);
@@ -51,7 +56,7 @@ public class ReviewService {
 
         // 상점 정보가 유효한지 확인
         MarketEntity market = marketRepository.findById(review.getStoreId())
-                .orElseThrow(()-> new RuntimeException("해당 상점을 찾을 수 없어 리뷰 등록 불가합니다. "));
+                .orElseThrow(()-> new RuntimeException("상점 정보를 찾을 수 없습니다. "));
 
         // 등록하고자 하는 상점을 이용한 이력이 있는지 확인
         List<Integer> reservationStates = new ArrayList<>();
@@ -73,32 +78,69 @@ public class ReviewService {
         newReview.setRating(review.getRating());
         reviewRepository.save(newReview);
         logger.info("end to create review");
-        return newReview;
 
+        // 해당 상점 정보를 찾아서 평균 상점 정보 업데이트
+        market.addAverRating(review.getRating() == null ? 0L : review.getRating(),
+                             market.getRatingCount() == null ? 0L : market.getRatingCount());
+
+
+
+        return newReview;
     }
 
     @Transactional
     public ReviewEntity updateReview(Review review){
         logger.info("started to update review");
-        ReviewEntity updateReview = reviewRepository.findById(review.getId())
+
+        ReviewEntity writtenReview = reviewRepository.findById(review.getId())
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 리뷰 글 입니다. "));
 
-        updateReview.setRegisterDt(review.getRegisterDt());
-        updateReview.setRegisterUserId(review.getRegisterUserId());
-        updateReview.setStoreId(review.getStoreId());
-        updateReview.setContent(review.getContent());
-        updateReview.setRating(review.getRating());
-        reviewRepository.save(updateReview);
+        // 상점 정보가 유효한지 확인
+        MarketEntity market = marketRepository.findById(writtenReview.getStoreId())
+                .orElseThrow(()-> new RuntimeException("상점 정보를 찾을 수 없습니다. "));
+
+        // 작성자가 수정하려고 하는지 확인
+        MemberEntity member = memberRepository.findById(writtenReview.getRegisterUserId())
+                .orElseThrow(() -> new RuntimeException("사용자 정보를 찾을 수 없습니다. [작성자만 수정가능합니다.]"));
+
+
+        // 기존 평점 삭제처리
+        market.deleteAverRating(writtenReview.getRating(), market.getRatingCount());
+
+        writtenReview.setContent(review.getContent());
+        writtenReview.setRating(review.getRating());
+        reviewRepository.save(writtenReview);
+
+        // 신규로 변경할 평점 정보 업데이트
+        market.addAverRating(writtenReview.getRating(), market.getRatingCount());
         logger.info("end to update market");
-        return updateReview;
+
+        return writtenReview;
     }
 
     @Transactional
     public ReviewEntity deleteReview(Long id) {
         logger.info("started to delete review");
+
         ReviewEntity delete = reviewRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 리뷰 글 입니다. "));
+
+        // 상점 정보가 유효한지 확인
+        MarketEntity market = marketRepository.findById(delete.getStoreId())
+                .orElseThrow(()-> new RuntimeException("상점 정보를 찾을 수 없습니다. "));
+
+
+        // 작성자 또는 관리자가 수정하려고 하는지 확인
+        MemberEntity member = memberRepository.findByUsername(userService.getCurrentUsername())
+                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+        if(!member.getId().equals(delete.getRegisterUserId()) && !member.getId().equals(market.getAdminId())) {
+            throw new RuntimeException("해당 게시물 삭제 권한이 없습니다." );
+        }
+
         reviewRepository.deleteById(delete.getId());
+        market.deleteAverRating(delete.getRating() == null ? 0L : delete.getRating(),
+                market.getRatingCount() == null ? 0L : market.getRatingCount());
+
         logger.info("end to delete review");
         return delete;
     }
